requirements.txt
a2wsgi==1.10.10
agate==1.9.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiologic==0.14.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-common-compat==1.7.1
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.2
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-google==16.0.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-sqlite==3.8.2
apache-airflow-providers-standard==1.3.0
apache-airflow-task-sdk==1.0.2
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backports.strenum==1.3.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
cadwyn==5.4.2
cattrs==25.1.1
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
colorama==0.4.6
colorful==0.5.7
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.4
daff==1.4.2
db-dtypes==1.4.3
dbt-adapters==1.16.0
dbt-bigquery==1.9.2
dbt-common==1.25.0
dbt-core==1.10.2
dbt-extractor==0.6.0
dbt-protos==1.0.330
dbt-semantic-interfaces==0.8.5
decorator==5.2.1
deepdiff==7.0.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.7.0
fsspec==2025.5.1
gcloud-aio-auth==5.4.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.4.0
gcsfs==2025.5.1
google-ads==27.0.0
google-analytics-admin==0.24.1
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.100.0
google-cloud-alloydb==0.4.8
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-automl==2.16.4
google-cloud-batch==0.17.36
google-cloud-bigquery==3.34.0
google-cloud-bigquery-datatransfer==3.19.2
google-cloud-bigquery-storage==2.32.0
google-cloud-bigtable==2.31.0
google-cloud-build==3.31.2
google-cloud-compute==1.31.0
google-cloud-container==2.57.0
google-cloud-core==2.4.3
google-cloud-datacatalog==3.27.1
google-cloud-dataflow-client==0.9.0
google-cloud-dataform==0.6.2
google-cloud-dataplex==2.10.2
google-cloud-dataproc==5.20.0
google-cloud-dataproc-metastore==1.18.3
google-cloud-dlp==3.31.0
google-cloud-kms==3.5.1
google-cloud-language==2.17.2
google-cloud-logging==3.12.1
google-cloud-managedkafka==0.1.11
google-cloud-memcache==1.12.2
google-cloud-monitoring==2.27.2
google-cloud-orchestration-airflow==1.17.5
google-cloud-os-login==2.17.2
google-cloud-pubsub==2.30.0
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-run==0.10.18
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.55.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.17.0
google-cloud-tasks==2.19.3
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.21.1
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-cloud-workflows==1.18.2
google-crc32c==1.7.1
google-genai==1.23.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.1
grpcio-gcp==0.2.2
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.7.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
lazy-object-proxy==1.11.0
leather==0.4.0
libcst==1.8.2
limits==5.4.0
linkify-it-py==2.0.3
litellm==1.73.6.post1
lockfile==0.12.2
looker-sdk==25.10.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mashumaro==3.14
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pandas-gbq==0.29.1
parsedatetime==2.6
pathspec==0.12.1
pendulum==2.1.2
platformdirs==4.3.8
pluggy==1.6.0
prison==0.2.1
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
ray==2.42.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.8
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.15.3
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.15.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.14.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-protobuf==6.30.2.20250703
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.0.1
yarl==1.20.1
zipp==3.23.0
